Move classes in heuristics.h/cpp to separate header/implementation files.
