+++ python/boomer/seco/cpp/heuristics/heuristic_hamming_loss.h
@@ -1,0 +1,24 @@
+/**
+ * @author Jakob Steeg (jakob.steeg@gmail.com)
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "heuristic.h"
+
+
+namespace seco {
+
+    /**
+     * A heuristic that measures the fraction of incorrectly predicted labels among all labels.
+     */
+    class HammingLoss : public IHeuristic {
+
+        public:
+
+            float64 evaluateConfusionMatrix(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin,
+                                            float64 uip, float64 urn, float64 urp) const override;
+
+    };
+
+}
