+++ python/boomer/seco/cpp/heuristics/heuristic_precision.cpp
@@ -1,0 +1,10 @@
+#include "heuristic_precision.h"
+#include "heuristic_common.h"
+
+using namespace seco;
+
+
+float64 Precision::evaluateConfusionMatrix(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin, float64 uip,
+                                           float64 urn, float64 urp) const {
+    return precision(cin, cip, crn, crp);
+}
