+++ python/boomer/seco/cpp/heuristics/heuristic_recall.cpp
@@ -1,0 +1,10 @@
+#include "heuristic_recall.h"
+#include "heuristic_common.h"
+
+using namespace seco;
+
+
+float64 Recall::evaluateConfusionMatrix(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin, float64 uip,
+                                        float64 urn, float64 urp) const {
+    return recall(cin, crp, uin, urp);
+}
