+++ python/boomer/seco/cpp/heuristics/heuristic_recall.h
@@ -1,0 +1,25 @@
+/**
+ * @author Jakob Steeg (jakob.steeg@gmail.com)
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "heuristic.h"
+
+
+namespace seco {
+
+    /**
+     * A heuristic that measures the fraction of uncovered labels among all labels for which the rule's prediction is
+     * (or would be) correct, i.e., for which the ground truth is equal to the rule's prediction.
+     */
+    class Recall : public IHeuristic {
+
+        public:
+
+            float64 evaluateConfusionMatrix(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin,
+                                            float64 uip, float64 urn, float64 urp) const override;
+
+    };
+
+}
