+++ python/boomer/seco/cpp/rule_evaluation_label_wise.h
@@ -10,7 +10,7 @@
 #include "../../common/cpp/indices/index_vector_full.h"
 #include "../../common/cpp/indices/index_vector_partial.h"
 #include "../../common/cpp/rule_evaluation.h"
-#include "heuristics.h"
+#include "heuristics/heuristic.h"
 #include <memory>
