+++ python/setup.py
@@ -66,7 +66,12 @@ sources = [
     'boomer/boosting/cpp/post_processing.cpp',
     'boomer/seco/cpp/head_refinement/head_refinement_partial.cpp',
     'boomer/seco/cpp/head_refinement/lift_function_peak.cpp',
-    'boomer/seco/cpp/heuristics.cpp',
+    'boomer/seco/cpp/heuristics/heuristic_f_measure.cpp',
+    'boomer/seco/cpp/heuristics/heuristic_hamming_loss.cpp',
+    'boomer/seco/cpp/heuristics/heuristic_m_estimate.cpp',
+    'boomer/seco/cpp/heuristics/heuristic_precision.cpp',
+    'boomer/seco/cpp/heuristics/heuristic_recall.cpp',
+    'boomer/seco/cpp/heuristics/heuristic_wra.cpp',
     'boomer/seco/cpp/statistics_label_wise.cpp',
     'boomer/seco/cpp/rule_evaluation_label_wise.cpp'
 ]
